map script now works locally, styling added to map
