+++ app/javascript/packs/map.js
@@ -1,0 +1,20 @@
+import mapboxgl from 'mapbox-gl';
+import 'mapbox-gl/dist/mapbox-gl.css';
+
+const mapElement = document.getElementById('map');
+
+if (mapElement) {
+  mapboxgl.accessToken = process.env.MAPBOX_API_KEY;
+  const map = new mapboxgl.Map({
+    container: 'map',
+    style: 'mapbox://styles/karinmiriam/cko88xcgl4vju18pgby8zpan6'
+    // style: 'mapbox://styles/mapbox/streets-v10'
+  });
+
+  const markers = JSON.parse(mapElement.dataset.markers);
+  markers.forEach((marker) => {
+  new mapboxgl.Marker()
+    .setLngLat([ marker.lng, marker.lat ])
+    .addTo(map);
+  });
+}
