+++ app/views/layouts/application.html.erb
@@ -8,9 +8,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= stylesheet_pack_tag 'map' %>
-    <%#= javascript_pack_tag 'map' %>
-    <script src='https://api.mapbox.com/mapbox-gl-js/v2.2.0/mapbox-gl.js'></script>
-
     <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload', defer: true %>
   </head>
 
@@ -22,8 +19,8 @@
       <%= yield %>
     </div>
 
-    <p class="notice"><%#= notice %></p>
-    <p class="alert"><%#= alert %></p>
+    <p class="notice"><%= notice %></p>
+    <p class="alert"><%= alert %></p>
 
 
     <div class="footer">
@@ -34,7 +31,7 @@
       </ul>
     </div>
 
-
+    <%= javascript_pack_tag 'map' %>
 
 
   </body>
